Add basic authentication
The client authenticates by sshing into the server and creating a token.
The server is now aware of the clients uid and pid. This pid must remain
constant for the duration of the user's session.

Part of #1. Messy, but okay for the time being. Will allow easy
verification of `cd` behaviour with regards to #2.