Update Door and add UnixFile class
Adds a new class UnixFile which represents a file (including
directory). Door can now be connected to a file node which it uses to
determine whether or not to let bodies through based on file
permissions.

Also added some labels to Door (only visible in editor) to make it
easier to determine which side is which.