Update SudoCan material per instance
The Mesh for SudoCan has a material Surface1 which can be
accessed by mesh.surface_get_material(0). When this material is modified
in updates for ALL instances.

Instead, we can override this material in MeshInstance by duplicating
the material and adding it to the same slot with
set_surface_material(0). That way, when the material is modified on a
mesh instance the changes to the material are only applied to that
instance.