Add world generation
Adds a new mesh_merger script to replace the godot-splerger addon that
can work with mirrored MeshInstances and updates the detour_navigation
script to use it.

Uses many ideas and algorithms from the game "Eldritch Reanimated",
which are discussed in the talk "Procedural Level Design in Eldritch"
available at https://www.youtube.com/watch?v=BYN0PJOdvzs
The games source code is available at https://www.eldritchgame.com/about.html#source
and is published under the following license:

/* ---------------------------------------------------------------- */

Copyright Š 2013-2019 Minor Key Games, LLC

This software is provided 'as-is', without any express or implied
warranty.  In no event will the authors be held liable for any damages
arising from the use of this software.

Permission is granted to anyone to use this software for any purpose,
including commercial applications, and to alter it and redistribute it
freely, subject to the following restrictions:

1. The origin of this software must not be misrepresented; you must not
   claim that you wrote the original software. If you use this software
   in a product, an acknowledgment in the product documentation would be
   appreciated but is not required.
2. Altered source versions must be plainly marked as such, and must not be
   misrepresented as being the original software.
3. This notice may not be removed or altered from any source distribution.

/* ---------------------------------------------------------------- */